display facet labels from scheming choices.
By default, CKAN displays the facet labels for custom fields as just the field value. CKAN does provide a function as an argument passed to snippets/facet_list called label_function allows the customization of the value label. This change update the search.html template to pass a custom labeler in if the field is part of the dataset scheming schema. the new function `saerischema_get_labeler_for_facet` calculates if the field is a scheming schema choices field. If so, a function is returned which will replace the value with the correct label. This works for both choices and choices_helper fields.